v1.09 updates
Fixed potential conflict with other addons (disambiguated Main.cs)
Created an option for enabling the Shadow out of Time cooldown bar. Enable with "/setoption alia_shadowbar true"
Note that this bar works pretty well for the first Shadow cast in phase 3 on E17 & SM, but is less reliable for 2nd and subsequent casts and may be less reliable on other difficulties. More data needed to nail down the timings.